Prevent delete own account

+++ src/actions/delete-user.ts
@@ -1,9 +1,20 @@
 'use server';
 
 import { db } from '@/lib/db';
+import { currentUser } from '@/lib/authentication';
 import { getErrorMessage } from '@/lib/handle-error';
 
 export async function deleteUser(user_ids: string[]) {
+  const user = await currentUser();
+
+  if (!user || !user.id) {
+    return { error: 'You are not authorized to delete users.' };
+  }
+
+  if (user_ids.includes(user.id)) {
+    return { error: 'You cannot delete your own account.' };
+  }
+
   try {
     await db.user.deleteMany({
       where: {
